Refactor hiding of tab bar and add SubtaskList styling
The layout effect function used in each stack to hide the tab bar was extracted into its own separate function to remove code duplication.

The SubtaskList has also had its styling improved, with better margin and padding along with updates to RectButton and the creation of ColumnContainer and RowContainer to make styling easier.